+++ packages/address-consts/README.md
@@ -1,0 +1,8 @@
+# `@shopify/address-consts`
+
+[![Build Status](https://travis-ci.org/Shopify/quilt.svg?branch=master)](https://travis-ci.org/Shopify/quilt)
+[![License: MIT](https://img.shields.io/badge/License-MIT-green.svg)](LICENSE.md) [![npm version](https://badge.fury.io/js/%40shopify%2Faddress-consts.svg)](https://badge.fury.io/js/%40shopify%2Faddress-consts.svg)
+
+Constants and types relating to `@shopify/address`.
+
+This package is not meant for direct consumption, rather it contains dependencies that both `@shopify/address` and `@shopify/address-mocks` require.
