+++ packages/address-consts/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "@shopify/address-consts",
+  "version": "0.0.1",
+  "license": "MIT",
+  "description": "Constants and types relating to @shopify/address",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "sideEffects": false,
+  "scripts": {
+    "build": "tsc --p tsconfig.build.json",
+    "prepublishOnly": "yarn run build"
+  },
+  "publishConfig": {
+    "access": "public",
+    "@shopify:registry": "https://registry.npmjs.org"
+  },
+  "author": "Shopify Inc.",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Shopify/quilt.git"
+  },
+  "bugs": {
+    "url": "https://github.com/shopify/quilt/issues"
+  },
+  "homepage": "https://github.com/Shopify/quilt/blob/master/packages/address-consts/README.md",
+  "dependencies": {},
+  "devDependencies": {
+    "typescript": "~3.2.1"
+  },
+  "files": ["dist/*"]
+}
