+++ packages/address/src/index.ts
@@ -1,3 +1,4 @@
-export * from './types';
+export * from '@shopify/address-consts';
+
 export * from './loader';
 export {default} from './AddressFormatter';
